refactoring semantics
