+++ src/test/scala/ai/privado/languageEngine/java/SemanticSecondLevelDerivationTest.scala
@@ -24,17 +24,19 @@
 package ai.privado.languageEngine.java
 
 import ai.privado.entrypoint.PrivadoInput
+import ai.privado.languageEngine.java.semantic.SemanticGenerator
 import ai.privado.languageEngine.java.tagger.source.{IdentifierTagger, InSensitiveCallTagger}
-import ai.privado.utility.Utilities
-import io.joern.dataflowengineoss.semanticsloader.FlowSemantic
+import io.joern.dataflowengineoss.semanticsloader.{FlowSemantic, Semantics}
 
 class SemanticSecondLevelDerivationTest extends JavaTaggingTestBase {
 
   val privadoScanConfig: PrivadoInput = PrivadoInput(disable2ndLevelClosure = true)
+  var semantics: Semantics            = Semantics.empty
   override def beforeAll(): Unit = {
     super.beforeAll()
     new IdentifierTagger(cpg, privadoScanConfig).createAndApply()
     new InSensitiveCallTagger(cpg).createAndApply()
+    semantics = SemanticGenerator.getSemantics(cpg, privadoScanConfig)
   }
   override val javaFileContents =
     """
@@ -80,33 +82,27 @@ class SemanticSecondLevelDerivationTest extends JavaTaggingTestBase {
 
   "Semantic generated for 2nd Level derivation" should {
     "have non-personal semantics for 2nd Level class by extends" in {
-      val semantics = Utilities.getSemantics(cpg, privadoScanConfig)
       semantics.elements.contains(FlowSemantic("User.getAmount:int()", List())) shouldBe true
       semantics.elements.contains(FlowSemantic("User.getId:java.lang.String()", List())) shouldBe true
     }
 
     "not have personal semantics for 2nd Level class by extends" in {
-      val semantics = Utilities.getSemantics(cpg, privadoScanConfig)
       semantics.elements.contains(FlowSemantic("User.getFirstName:int()", List())) shouldBe false
     }
 
     "have non-personal semantics for 1st Level class" in {
-      val semantics = Utilities.getSemantics(cpg, privadoScanConfig)
       semantics.elements.contains(FlowSemantic("BaseClass.getId:java.lang.String()", List())) shouldBe true
     }
 
     "not have personal semantics for 1st Level class" in {
-      val semantics = Utilities.getSemantics(cpg, privadoScanConfig)
       semantics.elements.contains(FlowSemantic("BaseClass.getFirstName:java.lang.String()", List())) shouldBe false
     }
 
     "have non-personal semantics for 2nd Level class by member" in {
-      val semantics = Utilities.getSemantics(cpg, privadoScanConfig)
       semantics.elements.contains(FlowSemantic("Customer.getType:java.lang.String()", List())) shouldBe true
     }
 
     "not have personal semantics for 2nd Level class by member" in {
-      val semantics = Utilities.getSemantics(cpg, privadoScanConfig)
       semantics.elements.contains(FlowSemantic("Customer.getBaseClass:BaseClass", List())) shouldBe false
     }
   }
